Binance.Api

Description

Contains a few useful functions for fetching data from the Binance API.

The main things to look at are getRates and getRatesMap for Exchange rates, and getAssets and getOrders for information about your account (these require a secret key).

All the functions that require both keys take the secret key first, and the regular key second.

This Module makes heavy use of Rationals for representing numbers given by the binance API. If you're gonna do a lot of computations on these numbers, converting them to a floating-point type using fromRational is recommended.

Documentation

data CoinInfo

Information about the amount of a coin that you posess.

coinFree is what you can directly use, and coinLocked is the amount reserved for orders. I don't know what coinFrozen does.

To get the total amount, you can use coinSum.

Constructors

CoinInfo

Fields

coinTicker :: String

coinFree :: Rational

coinFrozen :: Rational

coinLocked :: Rational

coinName :: String

getAssets :: ByteString -> ByteString -> IO (Either String [CoinInfo])

Uses your secret key and regular API key to fetch your assets:

>>> import Data.ByteString.Char8 (pack)
>>> getAssets (pack "private API key") (pack "regular API key")
Right [CoinInfo {coinTicker = "ETH", coinFree = 1 % 2, coinFree = 0 % 1, coinLocked = 0 % 1, coinName = "Ethereum"}]

This only fetches assets from your spot account, as the API for staked assets got added very recently, and I haven't implemented that API yet.

coinSum :: CoinInfo -> Rational

Sums up coinFree, coinFrozen and coinLocked for a CoinInfo

data ExchangeRate

Most recent exchange rate for a pair of coins. exSymbol has a value of the coins concatenated; This is not nice to work with, but the API only uses this format. For example, if an Ether was worth 2000 USDT, this would be represented as:

ExchangeRate {exSymbol = "ETHUSDT", exRate = 2000 % 1}

Constructors

ExchangeRate

Fields

exSymbol :: String

exRate :: Rational

getRates :: ByteString -> IO (Either String [ExchangeRate])

This fetches all the exchange rates, and returns them in a List. If you just want to look up specific exchange rates, getRatesMap is probably a better match for you.

getRatesMap :: ByteString -> IO (Either String (String -> String -> Either String Rational))

This fetches all the exchange rates on binance, and returns a function which you can use to look up pretty much all pairs of coins (if there's not an entry for a pair, the function will try to go though USDT before failing).

This builds a strict HashMap from getRates, and partially applies a lookup function that tries to indirectly look exchange rates up.

>>> import Data.ByteString.Char8 (pack)
>>> (Right getRate) <- getRatesMap $ pack "regular API key"
>>> fromRational <$> getRate "ETH" "USDT"
Right 2925.42
>>> fromRational <$> lookupRate "Blub" "Wubbel"
Left "Couldn't convert Blub -> Wubbel"

data OrderInfo

Represents an open order. For the meaning of non-self-explanatory fields see the binance API documentation.

A lot of fields are left out here, as I didn't need/understand them, but they're very easy to add, if you need them.

Constructors

OrderInfo

Fields

ordSymbol :: String

The pair, in the same format as exSymbol

ordPrice :: Rational

ordOrigQty :: Rational

ordExecutedQty :: Rational

ordStatus :: String

this should be an enum, but idk about possible values

ordType :: String

same for this field

ordSide :: String

and this field

ordTime :: UTCTime

getOrders :: ByteString -> ByteString -> IO (Either String [OrderInfo])

Uses your secret key and regular API key to fetch your open orders:

>>> import Data.ByteString.Char8 (pack)
>>> getOrders (pack "private API key") (pack "regular API key")
Right [OrderInfo {ordSymbol = "ALGOUSDT", ordPrice = 1 % 2, ordOrigQty = 22 % 1, ordExecutedQty = 0 % 1, ordStatus = "NEW", ordType = "LIMIT", ordSide = "SELL", ordTime = 2022-05-18 08:32:14 UTC}]